refactor(code_gen): extract methods/array.rs
